requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
Django==3.1.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.1
environs==8.0.0
factory-boy==3.1.0
Faker==4.14.0
idna==2.10
iniconfig==1.0.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.8.0
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.4
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
pytest-django==3.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
